backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "RAG backend with FastAPI, LangChain, Pinecone"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [ { name = "Windsurf" } ]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.2.16",
  "langchain-openai>=0.2.3",
  "pinecone>=5.0.1",
  "langchain-pinecone>=0.1.3",
  "langchain-community>=0.2.16",
  "langchain-text-splitters>=0.2.4",
  "tiktoken>=0.7.0",
  "httpx>=0.27.0",
  "orjson>=3.10.7",
  "tenacity>=9.0.0",
  "pypdf>=4.3.1",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.1",
  "ruff>=0.6.4",
  "black>=24.8.0",
  "types-requests",
]

[tool.uv]
# uv will read dependencies from [project]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","N"]
ignore = ["E203","E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.scripts]
app = "src.main:run"
